Продолжение заданий

+++ managercore/src/main/java/validation/classes/UserExistValidator.java
@@ -9,12 +9,12 @@ import javax.validation.ConstraintValidatorContext;
 /**
  * Валиадатор для проверки, что пользователь существует.
  */
-public class UserExistValidator implements ConstraintValidator<UserExist, Integer> {
+public class UserExistValidator implements ConstraintValidator<UserExist, Long> {
     @Override
     public void initialize(UserExist userExist){}
 
     @Override
-    public boolean isValid(Integer userId, ConstraintValidatorContext constraintValidatorContext) {
+    public boolean isValid(Long userId, ConstraintValidatorContext constraintValidatorContext) {
         //TODO првоерка в сервисе авторизации
         return true;
     }
